+++ mobile/android/android-components/components/lib/state/src/main/java/mozilla/components/lib/state/ext/StoreExtensions.kt
@@ -5,11 +5,11 @@
 package mozilla.components.lib.state.ext
 
 import android.view.View
+import androidx.annotation.MainThread
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleObserver
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.OnLifecycleEvent
-import androidx.lifecycle.ProcessLifecycleOwner
 import mozilla.components.lib.state.Action
 import mozilla.components.lib.state.Observer
 import mozilla.components.lib.state.State
@@ -20,8 +20,10 @@ import mozilla.components.lib.state.Store
  * will be bound to the passed in [LifecycleOwner]. Once the [Lifecycle] state changes to DESTROYED the [Observer] will
  * be unregistered automatically.
  *
- * Right after registering the [Observer] will be invoked with the current [State].
+ * The [Observer] will get invoked with the current [State] as soon as the [Lifecycle] is in STARTED
+ * state.
  */
+@MainThread
 fun <S : State, A : Action> Store<S, A>.observe(
     owner: LifecycleOwner,
     observer: Observer<S>
@@ -43,8 +45,16 @@ fun <S : State, A : Action> Store<S, A>.observe(
  * will be bound to the passed in [View]. Once the [View] gets detached the [Observer] will be unregistered
  * automatically.
  *
- * Right after registering the [Observer] will be invoked with the current [State].
+ * Note that inside a `Fragment` using [observe] with a `viewLifecycleOwner` may be a better option.
+ * Only use this implementation if you have only access to a [View] - especially if it can exist
+ * outside of a `Fragment`.
+ *
+ * The [Observer] will get invoked with the current [State] as soon as [View] is attached.
+ *
+ * Once the [View] gets detached the [Observer] will get unregistered. It will NOT get automatically
+ * registered again if the same [View] gets attached again.
  */
+@MainThread
 fun <S : State, A : Action> Store<S, A>.observe(
     view: View,
     observer: Observer<S>
@@ -53,6 +63,12 @@ fun <S : State, A : Action> Store<S, A>.observe(
 
     subscription.binding = SubscriptionViewBinding(view, subscription).apply {
         view.addOnAttachStateChangeListener(this)
+    }
+
+    if (view.isAttachedToWindow) {
+        // This View is already attached. We can resume immediately and do not need to wait for
+        // onViewAttachedToWindow() getting called.
+        subscription.resume()
     }
 }
 
@@ -64,7 +80,7 @@ fun <S : State, A : Action> Store<S, A>.observe(
 fun <S : State, A : Action> Store<S, A>.observeForever(
     observer: Observer<S>
 ) {
-    observe(ProcessLifecycleOwner.get(), observer)
+    observeManually(observer).resume()
 }
 
 /**
@@ -74,6 +90,16 @@ private class SubscriptionLifecycleBinding<S : State, A : Action>(
     private val owner: LifecycleOwner,
     private val subscription: Store.Subscription<S, A>
 ) : LifecycleObserver, Store.Subscription.Binding {
+    @OnLifecycleEvent(Lifecycle.Event.ON_START)
+    fun onStart() {
+        subscription.resume()
+    }
+
+    @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
+    fun onStop() {
+        subscription.pause()
+    }
+
     @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
     fun onDestroy() {
         subscription.unsubscribe()
@@ -91,7 +117,9 @@ private class SubscriptionViewBinding<S : State, A : Action>(
     private val view: View,
     private val subscription: Store.Subscription<S, A>
 ) : View.OnAttachStateChangeListener, Store.Subscription.Binding {
-    override fun onViewAttachedToWindow(v: View?) = Unit
+    override fun onViewAttachedToWindow(v: View?) {
+        subscription.resume()
+    }
 
     override fun onViewDetachedFromWindow(view: View) {
         subscription.unsubscribe()