+++ mobile/android/android-components/components/lib/state/src/test/java/mozilla/components/lib/state/ext/StoreExtensionsKtTest.kt
@@ -75,35 +75,39 @@ class StoreExtensionsKtTest {
             ::reducer
         )
 
+        // Observer does not get invoked since lifecycle is not started
         var stateObserved = false
         store.observe(owner) { stateObserved = true }
-        assertTrue(stateObserved)
+        assertFalse(stateObserved)
 
-        // CREATED
+        // CREATED: Observer does still not get invoked
         stateObserved = false
         owner.lifecycleRegistry.markState(Lifecycle.State.CREATED)
         store.dispatch(TestAction.IncrementAction).joinBlocking()
-        assertTrue(stateObserved)
+        assertFalse(stateObserved)
 
-        // STARTED
+        // STARTED: Observer gets initial state and observers updates
         stateObserved = false
         owner.lifecycleRegistry.markState(Lifecycle.State.STARTED)
+        assertTrue(stateObserved)
+
+        stateObserved = false
         store.dispatch(TestAction.IncrementAction).joinBlocking()
         assertTrue(stateObserved)
 
-        // RESUMED
+        // RESUMED: Observer continues to get updates
         stateObserved = false
         owner.lifecycleRegistry.markState(Lifecycle.State.RESUMED)
         store.dispatch(TestAction.IncrementAction).joinBlocking()
         assertTrue(stateObserved)
 
-        // CREATED
+        // CREATED: Not observing anymore
         stateObserved = false
         owner.lifecycleRegistry.markState(Lifecycle.State.CREATED)
         store.dispatch(TestAction.IncrementAction).joinBlocking()
-        assertTrue(stateObserved)
+        assertFalse(stateObserved)
 
-        // DESTROYED
+        // DESTROYED: Not observing
         stateObserved = false
         owner.lifecycleRegistry.markState(Lifecycle.State.DESTROYED)
         store.dispatch(TestAction.IncrementAction).joinBlocking()
@@ -148,6 +152,46 @@ class StoreExtensionsKtTest {
 
         stateObserved = false
         store.dispatch(TestAction.IncrementAction).joinBlocking()
+        assertTrue(stateObserved)
+
+        activity.windowManager.removeView(view)
+        assertFalse(view.isAttachedToWindow)
+
+        stateObserved = false
+        store.dispatch(TestAction.IncrementAction).joinBlocking()
+        assertFalse(stateObserved)
+    }
+
+    @Test
+    fun `Observer bound to view will not get notified about state changes until the view is attached`() {
+        val activity = Robolectric.buildActivity(Activity::class.java).create().get()
+        val view = View(testContext)
+
+        assertFalse(view.isAttachedToWindow)
+
+        val store = Store(
+            TestState(counter = 23),
+            ::reducer
+        )
+
+        var stateObserved = false
+        store.observe(view) { stateObserved = true }
+        assertFalse(stateObserved)
+
+        stateObserved = false
+        store.dispatch(TestAction.IncrementAction).joinBlocking()
+        assertFalse(stateObserved)
+
+        activity.windowManager.addView(view, WindowManager.LayoutParams(100, 100))
+        assertTrue(view.isAttachedToWindow)
+        assertTrue(stateObserved)
+
+        stateObserved = false
+        store.observe(view) { stateObserved = true }
+        assertTrue(stateObserved)
+
+        stateObserved = false
+        store.observe(view) { stateObserved = true }
         assertTrue(stateObserved)
 
         activity.windowManager.removeView(view)